Remove unused class diagram and renamed the ridesharingservice typo to match algomaster listing
