--- solutions/java/src/ridehailingservice/observer/TripObserver.java
@@ -1,7 +1,0 @@
-package ridehailingservice.observer;
-
-import ridehailingservice.entities.Trip;
-
-public interface TripObserver {
-    void onUpdate(Trip trip);
-}
