--- solutions/java/src/ridehailingservice/strategy/matching/DriverMatchingStrategy.java
@@ -1,11 +1,0 @@
-package ridehailingservice.strategy.matching;
-
-import ridehailingservice.entities.Driver;
-import ridehailingservice.entities.Location;
-import ridehailingservice.enums.RideType;
-
-import java.util.List;
-
-public interface DriverMatchingStrategy {
-    List<Driver> findDrivers(List<Driver> allDrivers, Location pickupLocation, RideType rideType);
-}
