+++ solutions/java/src/ridesharingservice/entities/User.java
@@ -1,10 +1,10 @@
-package ridehailingservice.entities;
-
-import ridehailingservice.observer.TripObserver;
+package ridesharingservice.entities;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+
+import ridesharingservice.observer.TripObserver;
 
 public abstract class User implements TripObserver {
     private final String id;
