+++ solutions/java/src/ridesharingservice/enums/TripStatus.java
@@ -1,4 +1,4 @@
-package ridehailingservice.enums;
+package ridesharingservice.enums;
 
 public enum TripStatus {
     REQUESTED,
