+++ solutions/java/src/ridesharingservice/observer/TripObserver.java
@@ -1,0 +1,7 @@
+package ridesharingservice.observer;
+
+import ridesharingservice.entities.Trip;
+
+public interface TripObserver {
+    void onUpdate(Trip trip);
+}
