+++ solutions/java/src/ridesharingservice/strategy/matching/NearestDriverMatchingStrategy.java
@@ -1,13 +1,13 @@
-package ridehailingservice.strategy.matching;
-
-import ridehailingservice.entities.Driver;
-import ridehailingservice.enums.DriverStatus;
-import ridehailingservice.entities.Location;
-import ridehailingservice.enums.RideType;
+package ridesharingservice.strategy.matching;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import ridesharingservice.entities.Driver;
+import ridesharingservice.entities.Location;
+import ridesharingservice.enums.DriverStatus;
+import ridesharingservice.enums.RideType;
 
 public class NearestDriverMatchingStrategy implements DriverMatchingStrategy {
     private static final double MAX_DISTANCE_KM = 5.0; // Max distance to consider a driver "nearby"
